Added Slideshow to the main page
The landing page will include a slideshow of venues around the city
along with a picture of the app describing something about it. Right now
the slideshow has been inserted but I plan on making it full screen
while the NiteMate and the download buttons appear in front of the
slideshow.